# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "osl_api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.28.0"
  },
  "paths": {
    "/api/v1/athletes": {
      "get": {
        "tags": [
          "athletes"
        ],
        "operationId": "list_athletes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of athletes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AthleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters"
          }
        }
      }
    },
    "/api/v1/athletes/{slug}": {
      "get": {
        "tags": [
          "athletes"
        ],
        "operationId": "get_athlete",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Athlete slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Include"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Athlete found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AthleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unknown include value"
          },
          "404": {
            "description": "Athlete not found"
          }
        }
      }
    },
    "/api/v1/competitions": {
      "get": {
        "tags": [
          "competitions"
        ],
        "operationId": "list_competitions",
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Include"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/CompetitionStatus"
                }
              ]
            }
          },
          {
            "name": "federation",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Direction"
            }
          },
          {
            "name": "pagination",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of competitions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CompetitionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination or include value"
          }
        }
      }
    },
...
    "/api/v1/competitions/{slug}": {
      "get": {
        "tags": [
          "competitions"
        ],
        "operationId": "get_competition",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Competition slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Include"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Competition found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompetitionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unknown include value"
          },
          "404": {
            "description": "Competition not found"
          }
        }
      }
    },
    "/api/v1/rankings": {
      "get": {
        "tags": [
          "rankings"
        ],
        "operationId": "get_global_ranking",
        "parameters": [
          {
            "name": "pagination",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationParams"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/RankedGender"
                }
              ]
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "federation",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "movement",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RankingMetric"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Direction"
            }
          },
          {
            "name": "event",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          },
          {
            "name": "competition_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Global ranking retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_GlobalRankingEntry"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          }
        }
      }
    },
    "/api/v1/rankings/classes": {
      "get": {
        "tags": [
          "rankings"
        ],
        "operationId": "list_ranking_classes",
        "parameters": [
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/Gender"
                }
              ]
            }
          },
          {
            "name": "competition_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Distinct weight classes, optionally narrowed to one gender",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rankings/countries": {
      "get": {
        "tags": [
          "rankings"
        ],
        "operationId": "list_ranking_countries",
        "parameters": [
          {
            "name": "competition_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Distinct countries athletes have competed under, optionally narrowed to one competition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ris/calculations": {
      "post": {
        "tags": [
          "ris"
        ],
        "operationId": "calculate_ris",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputeRisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "RIS computed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComputeRisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/api/v1/ris/formulas/{year}": {
      "get": {
        "tags": [
          "ris"
        ],
        "operationId": "get_formula_by_year",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "Edition year",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RIS edition for this year",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RisFormulaResponse"
                }
              }
            }
          },
          "404": {
            "description": "No edition was published for this year"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AthleteResponse": {
        "type": "object",
        "required": [
          "athlete_id",
          "first_name",
          "last_name",
          "slug",
          "gender",
          "country",
          "created_at"
        ],
        "properties": {
          "athlete_id": {
            "type": "string",
            "format": "uuid"
          },
          "competitions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/AthleteCompetitionSummary"
            }
          },
          "country": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "first_name": {
            "type": "string"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "instagram_handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_name": {
            "type": "string"
          },
          "native_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "personal_records": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/PersonalRecord"
            }
          },
          "profile_picture_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": "string"
          },
          "standing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AthleteStanding"
              }
            ]
          },
          "strength_profile": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StrengthProfile"
              }
            ]
          },
          "total_competitions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "CompetitionResponse": {
        "type": "object",
        "required": [
          "competition_id",
          "name",
          "created_at",
          "slug",
          "status",
          "federation_id"
        ],
        "properties": {
          "categories": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/CategoryDetail"
            }
          },
          "city": {
            "type": [
              "string",
              "null"
            ]
          },
          "competition_id": {
            "type": "string",
            "format": "uuid"
          },
          "country": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "end_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "event_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "federation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FederationInfo"
              }
            ]
          },
          "federation_id": {
            "type": "string",
            "format": "uuid"
          },
          "lifter_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "movements": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/MovementInfo"
            }
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": [
              "string",
              "null"
            ]
          },
          "slug": {
            "type": "string"
          },
          "start_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "status": {
            "$ref": "#/components/schemas/CompetitionStatus"
          }
        }
      },
      "CompetitionStatus": {
        "type": "string",
        "description": "Where a competition is in its life.",
        "enum": [
          "draft",
          "upcoming",
          "live",
          "completed",
          "cancelled"
        ]
      },
...
      "ComputeRisResponse": {
        "type": "object",
        "required": [
          "ris_score",
          "formula_year"
        ],
        "properties": {
          "formula_year": {
            "type": "integer",
            "format": "int32"
          },
          "ris_score": {
            "type": "string"
          }
        }
      },
...
      "Direction": {
        "type": "string",
        "description": "Which way a sorted list runs. Shared by the ranking and the competition\nlist, which disagree about the sensible default but not about the values.",
        "enum": [
          "desc",
          "asc"
        ]
      },
...
      "Gender": {
        "type": "string",
        "description": "Which category an athlete competes in.",
        "enum": [
          "M",
          "F",
          "MX"
        ]
      },
...
      "Include": {
        "type": "string",
        "description": "Comma-separated list of sections to embed"
      },
...
      "PaginatedResponse_AthleteResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "athlete_id",
                "first_name",
                "last_name",
                "slug",
                "gender",
                "country",
                "created_at"
              ],
              "properties": {
                "athlete_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "competitions": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/AthleteCompetitionSummary"
                  }
                },
                "country": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "first_name": {
                  "type": "string"
                },
                "gender": {
                  "$ref": "#/components/schemas/Gender"
                },
                "instagram_handle": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "last_name": {
                  "type": "string"
                },
                "native_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "personal_records": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/PersonalRecord"
                  }
                },
                "profile_picture_url": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "slug": {
                  "type": "string"
                },
                "standing": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/AthleteStanding"
                    }
                  ]
                },
                "strength_profile": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/StrengthProfile"
                    }
                  ]
                },
                "total_competitions": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                }
              }
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "PaginatedResponse_CompetitionResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "competition_id",
                "name",
                "created_at",
                "slug",
                "status",
                "federation_id"
              ],
              "properties": {
                "categories": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/CategoryDetail"
                  }
                },
                "city": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "competition_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "country": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "end_date": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date"
                },
                "event_code": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "federation": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/FederationInfo"
                    }
                  ]
                },
                "federation_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "lifter_count": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64"
                },
                "movements": {
                  "type": [
                    "array",
                    "null"
                  ],
                  "items": {
                    "$ref": "#/components/schemas/MovementInfo"
                  }
                },
                "name": {
                  "type": "string"
                },
                "region": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "slug": {
                  "type": "string"
                },
                "start_date": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date"
                },
                "status": {
                  "$ref": "#/components/schemas/CompetitionStatus"
                }
              }
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
      "PaginationParams": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "RankedGender": {
        "type": "string",
        "description": "The genders a ranking can be drawn for. Weight classes are only drawn for\nmen and women, so a mixed board would compare an athlete against an empty\nfield.",
        "enum": [
          "M",
          "F"
        ]
      },
      "RankingMetric": {
        "type": "string",
        "description": "What a ranking board is sorted by. Not only movements: a total and a RIS\nscore are ranked from the same list.",
        "enum": [
          "muscleup",
          "pullup",
          "dips",
          "squat",
          "total",
          "ris"
        ]
      },
...
    }
  },
...
}